separates jobs end points into individual router
