+++ backend/config/middleware.js
@@ -5,6 +5,7 @@ const helmet = require("helmet");
 const testRouter = require("../routers/testRouter.js");
 const router = require("../routers/router.js");
 const authRouter = require("../routers/authRouter.js");
+const jobsRouter = require("../routers/jobsRouter.js");
 
 module.exports = server => {
   // middleware
@@ -16,4 +17,5 @@ module.exports = server => {
   server.use("/test", testRouter);
   server.use("/api", router);
   server.use("/api/auth", authRouter);
+  server.use("/api/jobs", jobsRouter);
 };
